Moves database connection to `database.py` module
Tidies up some errors along the way
