+++ database.py
@@ -1,0 +1,42 @@
+"""Database connection"""
+
+from typing import AsyncGenerator
+
+from sqlalchemy.ext.asyncio import (
+    AsyncEngine,
+    AsyncSession,
+    async_sessionmaker,
+    create_async_engine,
+)
+
+from config import settings
+from models import Base
+
+async_engine: AsyncEngine = create_async_engine(url=settings.database_url)
+
+async_session: async_sessionmaker[AsyncSession] = async_sessionmaker(
+    bind=async_engine,
+    class_=AsyncSession,
+    expire_on_commit=False,
+)
+
+""" Create table"""
+
+
+async def create_tables() -> None:
+    async with async_engine.begin() as conn:
+        await conn.run_sync(Base.metadata.create_all)
+
+
+""" Yield sessions """
+
+
+async def get_db() -> AsyncGenerator[AsyncSession, None]:
+    async with async_session() as session:
+        try:
+            yield session
+        except Exception:
+            await session.rollback()
+            raise
+        finally:
+            await session.close()
