+++ main.py
@@ -1,33 +1,26 @@
-import asyncio
 import secrets
 import string
+from contextlib import asynccontextmanager
+from typing import AsyncGenerator
 
-from fastapi import FastAPI, HTTPException
+from fastapi import Depends, FastAPI, HTTPException
 from fastapi.responses import RedirectResponse
 from pydantic import BaseModel, HttpUrl, field_validator
-from sqlalchemy import Identity, Integer, String, select
-from sqlalchemy.dialects.postgresql import TEXT
-from sqlalchemy.ext.asyncio import (
-    AsyncEngine,
-    AsyncSession,
-    async_sessionmaker,
-    create_async_engine,
-)
-from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column
+from sqlalchemy import select
+from sqlalchemy.ext.asyncio import AsyncSession
 
 from config import settings
+from database import create_tables, get_db
 
 app: FastAPI = FastAPI(title=settings.app_name)
 
 
-""" Establish async connection to database """
-asyncio_engine: AsyncEngine = create_async_engine(url=settings.database_url)
+@asynccontextmanager
+async def lifespan(app: FastAPI) -> AsyncGenerator:
+    await create_tables()
+    yield
 
-async_session: async_sessionmaker[AsyncSession] = async_sessionmaker(
-    bind=asyncio_engine,
-    class_=AsyncSession,
-    expire_on_commit=False,
-)
+
 
 """SQLAlchemy models"""
 
@@ -54,16 +47,6 @@ class Link(Base):
             f"Link(link_id={self.link_id}, slug={self.slug}, long_url={self.long_url})"
         )
 
-
-""" Create table"""
-
-
-async def create_tables() -> None:
-    async with asyncio_engine.begin() as conn:
-        await conn.run_sync(Base.metadata.create_all)
-
-
-asyncio.run(create_tables())
 
 """ Custom errors """
 
@@ -162,22 +145,22 @@ async def read_root() -> dict:
 
 
 @app.post("/shorten")
-async def return_short_url(payload: LongUrlAccept) -> ShortUrlReturn:
-    async with async_session() as session:
-        try:
-            short_url: str = await create_short_url(
-                db=session, long_url=str(payload.long_url)
-            )
-            return ShortUrlReturn(short_url=short_url)  # type: ignore
-        except Exception as e:
-            raise HTTPException(status_code=500, detail="Internal error") from e
+async def return_short_url(
+    payload: LongUrlAccept, db: AsyncSession = Depends(get_db)
+) -> ShortUrlReturn:
+    try:
+        short_url: str = await create_short_url(db=db, long_url=str(payload.long_url))
+        return ShortUrlReturn(short_url=short_url)  # type: ignore
+    except Exception as e:
+        raise HTTPException(status_code=500, detail="Internal error") from e
 
 
 @app.get("/{slug}")
-async def return_long_url(slug: str) -> RedirectResponse:
-    async with async_session() as session:
-        try:
-            long_url: str | None = await get_long_url(db=session, slug=slug)
-            return RedirectResponse(long_url)
-        except NoMatchingSlugError as e:
-            raise HTTPException(status_code=404, detail="No link found") from e
+async def return_long_url(
+    slug: str, db: AsyncSession = Depends(get_db)
+) -> RedirectResponse:
+    try:
+        long_url: str | None = await get_long_url(db=db, slug=slug)
+        return RedirectResponse(long_url)
+    except NoMatchingSlugError as e:
+        raise HTTPException(status_code=404, detail="No link found") from e
